Refactor pages into separate modules
